pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mcph = "mcph.__main__:run"


[tool.poetry]
name = "mcph"
description = "Minecraft plugin helper, updates and checks versions of all plugins on a server!"
version = "0.1.0-beta.dev"
license = "AGPL3"

authors = ["PerchunPak <perchunpak@gmail.com>"]

readme = "README.md"

repository = "https://github.com/PerchunPak/mcph"

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Home Automation",
  "Topic :: Internet",
  "Topic :: Internet :: File Transfer Protocol (FTP)",
  "Topic :: Software Development :: Version Control",
]

[tool.poetry.dependencies]
python = "^3.8"

# Our dependecies
pyyaml = "6.0"
requests = ">=2.13"
prettytable = "~3.4"
typer = "~0.6"
omegaconf = "~2.2"

structlog = "21.5.0"
### For colors in structlog ###
### Only needed for Windows ###
colorama = { version = "*", markers="platform_system == 'Windows'" }

# Docs
sphinx                   = { version = "~5.0",  optional = true }
sphinx-autodoc-typehints = { version = "~1.18", optional = true }
sphinx-autoapi           = { version = ">=1.8,<1.10",  optional = true }
furo                     = { version = "*",     optional = true }
m2r2                     = { version = "~0.3",  optional = true }
tomli                    = { version = "~2.0",  optional = true }

[tool.poetry.dev-dependencies]
black = "~22"
isort = "~5.10"
pycln = "~2.0"
doc8 = "~0.11"
pre-commit = "~2.19"
mypy = "0.961"
flake8 = "~4"
flake8-docstrings = "~1"
pydocstyle = ">=4.0.0"

safety = "~2.0"

pytest = "~7.1"
pytest-cov = "~3.0"
pytest-mock = "~3.8"
pytest-testmon = "~1.3"
pytest-randomly = "~3.12"

[tool.poetry.extras]
docs = ["sphinx", "sphinx-autodoc-typehints", "sphinx-autoapi", "furo", "m2r2", "tomli"]

[tool.black]
line-length = 120
target-version = ['py310']
